# docs/swagger.yaml
definitions:
  api.CreateCourseRequest:
    properties:
      description:
        type: string
      image:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        type: string
      user_id:
        type: integer
    required:
    - description
    - title
    - type
    type: object
  api.CreateEnrolmentsRequest:
    properties:
      course_id:
        type: integer
      request_id:
        type: integer
      user_id:
        type: integer
    type: object
  api.CreateMaterialRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      title:
        type: string
    required:
    - course_id
    - description
    - title
    type: object
  api.ListUserRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
  api.ListUserStudentRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
  api.ListUserTeacherRequest:
    properties:
      page_id:
        minimum: 1
        type: integer
      page_size:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
  api.UpdateAssignmentRequest:
    properties:
      assignment_id:
        type: integer
      description:
        type: string
      submission_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  api.UpdateCoursesRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      image:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  api.UpdateMaterialRequest:
    properties:
      course_id:
        type: integer
      description:
        type: string
      material_id:
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - course_id
    - description
    - material_id
    - title
    type: object
  api.UpdateUserRequest:
    type: object
  api.createAssignmentRequest:
    properties:
      description:
        type: string
      resource_id:
        type: integer
      submission_date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  api.createCourseProgressRequest:
    properties:
      enrolment_id:
        type: integer
      progress:
        type: string
    type: object
  api.createRequestRequest:
    properties:
      course_id:
        type: integer
      is_accepted:
        type: boolean
      is_active:
        type: boolean
      is_declined:
        type: boolean
      is_pending:
        type: boolean
      user_id:
        type: integer
    type: object
  api.createResourceRequest:
    properties:
      content_url:
        type: string
      material_id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  api.createUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      hashed_password:
        type: string
      qualification:
        type: string
      role:
        type: string
      user_name:
        type: string
    type: object
  api.loginUserRequest:
    properties:
      hashed_password:
        type: string
      user_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Create a new admin's user with the provided details
      operationId: create admin user
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a admin's user
  /assignments:
    post:
      consumes:
      - application/json
      description: Create a new assignment with the provided details
      operationId: create-assignment
      parameters:
      - description: Assignment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new assignment
  /assignments/{assignment_id}:
    put:
      consumes:
      - application/json
      description: Update an assignment with the provided details
      operationId: update-assignment
      parameters:
      - description: Updated assignment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update an assignment
  /assignments/{assignment_id}/{resource_id}:
    delete:
      consumes:
      - application/json
      description: Delete an assignment by its ID and resource ID
      operationId: delete-assignment
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete an assignment
    get:
      consumes:
      - application/json
      description: Get an assignment by its ID and resource ID
      operationId: get-assignment
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get an assignment by ID
  /course:
    post:
      consumes:
      - application/json
      description: Creates a new course
      parameters:
      - description: Course details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new course
  /course/{course_id}:
    delete:
      description: Deletes a course by ID
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      - description: Teacher ID
        in: query
        name: teacher_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a course
    get:
      description: Retrieves a course by its ID
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a course by ID
  /course/update:
    put:
      consumes:
      - application/json
      description: Updates a course with provided details
      parameters:
      - description: Updated course details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCoursesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a course
  /courseEnrolments:
    get:
      consumes:
      - application/json
      description: List enrolments for a student in a course
      operationId: list-enrolments
      parameters:
      - description: Student ID
        in: query
        name: student_id
        required: true
        type: integer
      - description: Course ID
        in: query
        name: course_id
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List enrolments
  /courseProgress:
    get:
      consumes:
      - application/json
      description: Returns a list of course progress
      parameters:
      - description: Enrolment ID
        in: query
        name: enrolment_id
        required: true
        type: integer
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Offset for paginated results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List course progress
      tags:
      - CourseProgress
    post:
      consumes:
      - application/json
      description: Creates course progress
      parameters:
      - description: Course progress object
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/api.createCourseProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create course progress
      tags:
      - CourseProgress
  /courseProgress/{courseprogress_id}:
    get:
      consumes:
      - application/json
      description: Returns course progress by ID
      parameters:
      - description: Course progress ID
        in: path
        name: courseprogress_id
        required: true
        type: integer
      - description: Enrolment ID
        in: query
        name: enrolment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get course progress by ID
      tags:
      - CourseProgress
  /courses:
    get:
      description: Lists courses with pagination
      parameters:
      - description: Number of items to return
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List courses
  /enrolments:
    post:
      consumes:
      - application/json
      description: Create a new enrolment with the provided details
      operationId: create-enrolment
      parameters:
      - description: enrolment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateEnrolmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new enrolment
  /listadmin:
    get:
      consumes:
      - application/json
      description: ListUser with the provided admin based
      operationId: list-user
      parameters:
      - description: admin list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUser
  /liststudent:
    get:
      consumes:
      - application/json
      description: ListUserStudent with the provided student based
      operationId: list-student
      parameters:
      - description: student list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUserStudent
  /listteacher:
    get:
      consumes:
      - application/json
      description: ListUserTeacher with the provided teacher based
      operationId: list-teacher
      parameters:
      - description: teacher list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ListUserTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: ListUserTeacher
  /login:
    post:
      consumes:
      - application/json
      description: Log in a user with the provided credentials
      operationId: login-user
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Log in user
  /material:
    post:
      consumes:
      - application/json
      description: Create a new material
      operationId: create-material
      parameters:
      - description: Create Material Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new material
  /material/{course_id}:
    get:
      consumes:
      - application/json
      description: Get materials for a course
      operationId: get-materials
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get materials for a course
  /material/{material_id}:
    delete:
      consumes:
      - application/json
      description: Delete a material
      operationId: delete-material
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a material
    put:
      consumes:
      - application/json
      description: Update a material
      operationId: update-material
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      - description: Update Material Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a material
  /request/{request_id}:
    get:
      description: Get a request by its ID
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a request by ID
  /request/{user_id}/{request_id}:
    delete:
      description: Delete a request by student and request ID
      parameters:
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: integer
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a request
...
  /requests:
    get:
      description: List requests based on provided parameters
      parameters:
      - description: Student ID
        in: query
        name: student_id
        type: integer
      - description: Teacher ID
        in: query
        name: teacher_id
        type: integer
      - description: Course ID
        in: query
        name: course_id
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List requests
    post:
      consumes:
      - application/json
      description: Create a new request with the given parameters
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new request
  /resource:
    post:
      consumes:
      - application/json
      description: Creates a new resource
      operationId: create-resource
      parameters:
      - description: Create Resource Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new resource
  /resource/{resource_id}:
    delete:
      description: Deletes a resource
      operationId: delete-resource
      parameters:
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: integer
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a resource
...
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      operationId: create-user
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create a new user
  /submissions:
    get:
      description: Lists submissions for a given assignment ID
      operationId: listSubmissions
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List submissions
  /submissions/{assignment_id}/{student_id}:
    get:
      description: Retrieves a submission by assignment and student ID
      operationId: getSubmission
      parameters:
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a submission
  /tokens/renew_access:
    post:
      consumes:
      - application/json
      description: Renew access token using refresh token
      operationId: renew-access-token
      parameters:
      - description: Refresh Token
        in: body
        name: refresh_token
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Renew Access Token
  /user/edit:
    patch:
      consumes:
      - application/json
      description: Updates a user with provided details
      parameters:
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a user
swagger: "2.0"